# src/main/resources/static/openapi.yml
openapi: 3.1.0
info:
  title: Computer Architecture Course Administration API
  description: This API provides backend functionality for managing the Computer Architecture course system.
#  termsOfService: ?
#  contact:
#    name: ?
#    email: ?
#    url: ?
#  license:
#    name: ?
#    url: ?
  version: 1.0.0
#externalDocs:
#  description: ?
#  url: ?
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/block:
    get:
      tags:
        - BlockController
      summary: Retrieve a list of blocks
      description: >
        This endpoint allows you to fetch a list of blocks with optional sorting and filtering.

        - **Sorting** is available for all fields defined in `BlockRequestDTO` (including `id`).

        - **Filtering** is available for `id` (can be used at the same time).
      operationId: getBlocksByCriteria
      parameters:
        - $ref: '#/components/parameters/sort'
        - name: id
          in: query
          description: Filter blocks by ID.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - List of blocks returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlockResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    post:
      tags:
        - BlockController
      summary: Create a new block/blocks
      description: Adds a new block/blocks to the system with the provided details.
      operationId: addBlocks
      requestBody:
        description: Array of block objects that need to be created.
        content:
          application/json:
            schema:
              type: object
              properties:
                blocks:
                  type: array
                  items:
                    $ref: '#/components/schemas/BlockRequestDTO'
        required: true
      responses:
        '201':
          description: Created - Blocks created successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    put:
      tags:
        - BlockController
      summary: Update block by ID
      description: Updates information for an existing block identified by its ID.
      operationId: updateBlockById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the block.
          schema:
            type: integer
            format: int32
      requestBody:
        description: Block object with updated fields.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockRequestDTO'
        required: true
      responses:
        '200':
          description: OK - Block updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Block not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    delete:
      tags:
        - BlockController
      summary: Delete block by ID
      description: Removes a block from the system identified by its ID.
      operationId: deleteBlockById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the block.
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - Block deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Block not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
  /api/v1/assignment:
    get:
      tags:
        - AssignmentController
      summary: Retrieve a list of assignments
      description: >
        This endpoint allows you to fetch a list of assignments with optional sorting and filtering.

        - **Sorting** is available for all fields defined in `AssignmentRequestDTO` (including `id`).

        - **Filtering** is available for `id` (can be used at the same time).
      operationId: getAssignmentsByCriteria
      parameters:
        - $ref: '#/components/parameters/sort'
        - name: id
          in: query
          description: Filter assignments by ID.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - List of assignments returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssignmentResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    post:
      tags:
        - AssignmentController
      summary: Create a new assignment/assignments
      description: Adds a new assignment/assignments to the system with the provided details.
      operationId: addAssignments
      requestBody:
        description: Array of assignment objects that need to be created.
        content:
          application/json:
            schema:
              type: object
              properties:
                assignments:
                  type: array
                  items:
                    $ref: '#/components/schemas/AssignmentRequestDTO'
        required: true
      responses:
        '201':
          description: Created - Assignments created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Block not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    put:
      tags:
        - AssignmentController
      summary: Update assignment by ID
      description: Updates information for an existing assignment identified by its ID.
      operationId: updateAssignmentById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the assignment.
          schema:
            type: integer
            format: int32
      requestBody:
        description: Assignment object with updated fields.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentRequestDTO'
        required: true
      responses:
        '200':
          description: OK - Assignment updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Assignment or Block not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    delete:
      tags:
        - AssignmentController
      summary: Delete assignment by ID
      description: Removes an assignment from the system identified by its ID.
      operationId: deleteAssignmentById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the assignment.
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - Assignment deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Assignment not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
components:
  parameters:
    sort:
      name: sort
      in: query
      required: false
      description: >
        Sort the results by none or any field from the corresponding `Request/ResponseDTO`.
        Format: `field,(asc|desc)` >>> for example: `id,asc` or `id,desc`.
        If no order is specified, default is ascending.  
        Multiple sort parameters can be included, like:  
        `sort=field1,asc&sort=field2,desc`
      schema:
        type: string
  schemas:
...
    ErrorResponseDTO:
      type: object
      description: Represents a standard error response containing status code, general message, and optional list of detailed error messages.
      required:
        - status
        - message
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code representing the error.
        message:
          type: string
          description: A short description or summary of the error.
        errors:
          type: array
          items:
            type: string
          description: A list of detailed error messages.
...
    BlockRequestDTO:
      type: object
      description: Request body used for creating new block or updating existing block.
      required:
        - name
        - maxPoints
        - requiredPoints
      properties:
        name:
          type: string
          description: The block's name.
        maxPoints:
          type: number
          format: double
          description: The maximum number of points for the block.
        requiredPoints:
          type: number
          format: double
          description: The required points to pass the block.
...
    AssignmentRequestDTO:
      type: object
      description: Request body used for creating new assignment or updating existing assignment.
      required:
        - blockId
        - name
        - maxPoints
      properties:
        blockId:
          type: integer
          format: int32
          description: The unique identifier for the block associated with the assignment.
        name:
          type: string
          description: The assignment's name.
        maxPoints:
          type: number
          format: double
          description: The maximum number of points for the assignment.
...
    SuccessResponseDTO:
      type: object
      description: Represents a standard success response containing status code, general message, and optional list of data.
      required:
        - status
        - message
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code representing the success.
        message:
          type: string
          description: A short description or summary of the success.
        data:
          type: array
          items:
            type: object
          description: A list of updated or deleted data.
...
    BlockResponseDTO:
      type: object
      description: Response body that returns details about a block.
      required:
        - id
        - name
        - maxPoints
        - requiredPoints
      properties:
        id:
          type: integer
          format: int32
          description: The block's id.
        name:
          type: string
          description: The block's name.
        maxPoints:
          type: number
          format: double
          description: The maximum number of points the block can have.
        requiredPoints:
          type: number
          format: double
          description: The required points to pass the block, which must be less than or equal to maxPoints.
...
    AssignmentResponseDTO:
      type: object
      description: Response body that returns details about an assignment.
      required:
        - id
        - block
        - name
        - maxPoints
      properties:
        id:
          type: integer
          format: int32
          description: The assignment's id.
        block:
          type: object
          description: Details about the block to which the assignment belongs.
        name:
          type: string
          description: The assignment's name.
        maxPoints:
          type: number
          format: double
          description: The maximum number of points the assignment can have.
...